flytt del-med-bruker til modia-app
