+++ frontend/mulighetsrommet-veileder-flate/src/apps/modia/delMedBruker/Delemodal.module.scss
@@ -36,3 +36,8 @@
     margin-top: 1rem;
   }
 }
+
+.knapperad {
+  display: flex;
+  gap: 2rem;
+}
