+++ frontend/mulighetsrommet-veileder-flate/src/apps/modia/delMedBruker/Delemodal.tsx
@@ -4,17 +4,16 @@ import {
   DelMedBruker,
   VeilederflateTiltaksgjennomforing,
 } from "mulighetsrommet-api-client";
-import { PORTEN } from "../../../../../frontend-common/constants";
-import { mulighetsrommetClient } from "../../../core/api/clients";
-import { useLogEvent } from "../../../logging/amplitude";
-import { byttTilDialogFlate } from "../../../utils/DialogFlateUtils";
-import { erPreview } from "../../../utils/Utils";
-import modalStyles from "../../modal/Modal.module.scss";
-import { StatusModal } from "../../modal/StatusModal";
+import { PORTEN } from "mulighetsrommet-frontend-common/constants";
+import { mulighetsrommetClient } from "@/core/api/clients";
+import { useLogEvent } from "@/logging/amplitude";
+import { byttTilDialogFlate } from "@/utils/DialogFlateUtils";
+import { erPreview } from "@/utils/Utils";
+import { StatusModal } from "@/components/modal/StatusModal";
 import { DelMedBrukerContent, MAKS_ANTALL_TEGN_DEL_MED_BRUKER } from "./DelMedBrukerContent";
 import delemodalStyles from "./Delemodal.module.scss";
 import { Actions, State } from "./DelemodalActions";
-import { erBrukerReservertMotElektroniskKommunikasjon } from "../../../utils/Bruker";
+import { erBrukerReservertMotElektroniskKommunikasjon } from "@/apps/modia/delMedBruker/helpers";
 
 interface DelemodalProps {
   brukernavn?: string;
@@ -160,7 +159,7 @@ export function Delemodal({
             </HStack>
           </Modal.Body>
           <Modal.Footer>
-            <div className={modalStyles.knapperad}>
+            <div className={delemodalStyles.knapperad}>
               <Button
                 variant="tertiary"
                 onClick={clickCancel}
