+++ frontend/mulighetsrommet-veileder-flate/src/layouts/ViewTiltaksgjennomforingDetaljer.module.scss
@@ -40,10 +40,6 @@
   display: flex;
 }
 
-.deleknapp {
-  width: 100%;
-}
-
 .sidemeny {
   height: fit-content;
   margin-top: 3rem;
